requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
pillow==10.4.0
pytubefix==6.13.0
requests==2.32.3
urllib3==2.2.2
